Initial version of refactored S3D code
This is a first attempt at refactoring the S3D profile code into a more modular approach. The next step after this would be to integrate the current cura code into the same data models, so that the only difference between S3D and cura is in the templates used.

This introduces a new requirement, using Jinja to render the output files.

This is still work in progress, there are several open questions and bugs in the output.